NodeSerializer unserialize
